refactor(agent-runner): split session.rs into directory module to pass 600-line cap
session.rs grew to 737 lines after adding the natural-exit handler.
Extract spawn_natural_exit_handler into session/natural_exit.rs (154 lines).
session/mod.rs is now 596 lines, session/tests.rs 405 lines.
No logic changes — build, tests (8/8), clippy, and fmt all pass.
