+++ services/agent-runner/src/session/mod.rs
@@ -17,6 +17,8 @@ use tracing::Instrument;
 
 use crate::adapters::{AgentProcess, RuntimeAdapter, SessionCtx, adapter_for_runtime};
 
+mod natural_exit;
+
 const PROCESS_TERMINATE_TIMEOUT_SECS: u64 = 30;
 const MAX_INITIAL_PROMPT_LEN: usize = 100_000;
 const MAX_TRACKED_SESSIONS: usize = 100_000;
@@ -63,149 +65,6 @@ fn safe_join(base: &std::path::Path, rel: &str) -> Result<PathBuf> {
         }
     }
     Ok(base.join(rel_path))
-}
-
-/// Spawns a background task that waits for the child process to exit naturally
-/// and then transitions the session to `terminated` (exit 0) or `failed`
-/// (non-zero exit).
-///
-/// Race safety: the task removes the session from `sessions` as its first
-/// write action.  `terminate_session` also removes from `sessions` first.
-/// Whichever removes first owns the cleanup; the other returns without action.
-#[allow(clippy::too_many_arguments, clippy::too_many_lines)]
-fn spawn_natural_exit_handler(
-    session_id: String,
-    tenant_id: TenantId,
-    process: Arc<Mutex<AgentProcess>>,
-    sessions: Arc<Mutex<HashMap<String, SessionHandle>>>,
-    seq_counters: Arc<Mutex<HashMap<String, i64>>>,
-    seq_timestamps: Arc<Mutex<HashMap<String, Instant>>>,
-    control_api_base: String,
-    http_client: reqwest::Client,
-    event_sender: tokio::sync::mpsc::Sender<(TenantId, AgentEvent)>,
-) -> JoinHandle<()> {
-    let span = tracing::info_span!("natural_exit_handler", session_id = %session_id);
-
-    tokio::spawn(
-        async move {
-            // Wait for the process to exit naturally (releases lock before HTTP calls).
-            let exit_status = {
-                let mut proc = process.lock().await;
-                proc.child.wait().await
-            };
-
-            // Race check: try to claim ownership of cleanup by removing from the
-            // sessions map.  If terminate_session already removed it, bail out.
-            let Some(handle) = ({
-                let mut map = sessions.lock().await;
-                map.remove(&session_id)
-            }) else {
-                tracing::debug!(
-                    session_id = %session_id,
-                    "natural exit: session already removed by explicit terminate"
-                );
-                return;
-            };
-
-            // We own the cleanup.  Abort I/O handlers (wait_handle = self; dropping
-            // it does not abort in tokio, so no self-abort risk).
-            handle.stdout_handle.abort();
-            handle.stderr_handle.abort();
-
-            // Clean up seq-counter state.
-            {
-                let mut counters = seq_counters.lock().await;
-                let mut timestamps = seq_timestamps.lock().await;
-                counters.remove(&session_id);
-                timestamps.remove(&session_id);
-            }
-
-            let exit_code = match exit_status {
-                Ok(status) => status.code().unwrap_or(-1),
-                Err(_) => -1,
-            };
-            let duration_ms =
-                i64::try_from(handle.start_time.elapsed().as_millis()).unwrap_or(i64::MAX);
-            let final_status = if exit_code == 0 {
-                "terminated"
-            } else {
-                "failed"
-            };
-            let error_msg =
-                (exit_code != 0).then(|| format!("process exited with code {exit_code}"));
-
-            let Ok(validated_id) = uuid::Uuid::parse_str(&session_id) else {
-                tracing::warn!(
-                    session_id = %session_id,
-                    "natural exit: rejected non-UUID session_id"
-                );
-                return;
-            };
-
-            // Update session status in control-api.
-            let status_url =
-                format!("{control_api_base}/internal/agent/sessions/{validated_id}/status");
-            let body = serde_json::json!({
-                "status": final_status,
-                "pid": serde_json::Value::Null,
-                "exit_code": exit_code,
-                "error": error_msg,
-                "tenant_id": tenant_id.to_string(),
-            });
-            if let Err(e) = http_client.patch(&status_url).json(&body).send().await {
-                tracing::warn!(
-                    session_id = %session_id,
-                    "natural exit: failed to update session status: {e}"
-                );
-            }
-
-            // Revoke session-scoped API key.
-            let revoke_url =
-                format!("{control_api_base}/internal/agent/sessions/{validated_id}/api-key");
-            if let Err(e) = http_client.delete(&revoke_url).send().await {
-                tracing::warn!(
-                    session_id = %session_id,
-                    "natural exit: failed to revoke API key: {e}"
-                );
-            }
-
-            // Emit Terminated lifecycle event so the SSE stream closes cleanly.
-            let payload = serde_json::json!({
-                "exit_code": exit_code,
-                "duration_ms": duration_ms,
-                "reason": "natural_exit",
-            });
-            let event = AgentEvent {
-                tenant_id: tenant_id.to_string(),
-                session_id: session_id.clone(),
-                seq: TERMINATED_SEQ,
-                kind: AgentEventKind::Terminated.into(),
-                payload: payload.to_string(),
-                emitted_at_ms: chrono::Utc::now().timestamp_millis(),
-            };
-            if tokio::time::timeout(
-                Duration::from_secs(5),
-                event_sender.send((tenant_id, event)),
-            )
-            .await
-            .is_err()
-            {
-                tracing::warn!(
-                    session_id = %session_id,
-                    "Event channel full, dropped natural exit event"
-                );
-                counter!("rb_agent_events_dropped_total", "reason" => "channel_full").increment(1);
-            }
-
-            tracing::info!(
-                session_id = %session_id,
-                exit_code = exit_code,
-                duration_ms = duration_ms,
-                "Session completed naturally"
-            );
-        }
-        .instrument(span),
-    )
 }
 
 impl SessionManager {
@@ -355,7 +214,7 @@ impl SessionManager {
         let process_arc = Arc::new(Mutex::new(process));
         let start_time = Instant::now();
 
-        let wait_handle = spawn_natural_exit_handler(
+        let wait_handle = natural_exit::spawn_natural_exit_handler(
             session_id.to_string(),
             tenant_id,
             Arc::clone(&process_arc),
@@ -733,5 +592,5 @@ impl SessionManager {
 pub use crate::workspace_gc::spawn_workspace_gc;
 
 #[cfg(test)]
-#[path = "session_tests.rs"]
+#[path = "tests.rs"]
 mod tests;
